PointTree: Added a TreeADT, I believe this will be a more effective way of storing and traversing a list of points, easier for us to visualise too, as a bio-morph visually looks like a point tree.
BioController and BioGeneration:
Just added some lines of code to test the genes work well as I changed
the way they were stored within the gene class.

Gene:
Genes are now created in a for loop, currently (i<3) so this will create
3 genes, you can just change that number to generate any number of genes
we are going to use in the future. 

Point:
Filled in the skeleton code for this class, you are now able to
construct a point that consists of two integers, x and y.