Genes now generated using seed
Seed is generated and set as a variable in Biomorph. No longer passed
into Gene as this would reset the seed every time a new Gene is created.

Factory class gene generation methods no longer take the Seed as a
parameter. The gene values are calculated straight after setting the
seed in Biomorph.

Separate constructor for Genes that are branches.

Cleaned up some code in Biomorph class.